arreglo migration

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('avatar');
-            $table->numeric('score');
+            $table->integer('score');
             $table->boolean('isAdmin');
             $table->boolean('isActive');
             $table->rememberToken();
